+++ test/test.ts
@@ -7,23 +7,34 @@ import { JPCMConfig } from '../src/config';
 interface CommitMessageToTest {
   initialMessage: string[];
   expectedMessage: string;
-  flags?: string;
+  options?: {
+    isBranchIgnored: boolean;
+  };
   config?: Partial<JPCMConfig>;
 }
 
 const singleScopeMessage: CommitMessageToTest = {
   initialMessage: ['chore(deps): Finally solved that problem!'],
   expectedMessage: 'chore(deps): [JIRA-4321]. Finally solved that problem!',
+  config: {
+    isConventionalCommit: true,
+  },
 };
 
 const hyphenatedScopeMessage: CommitMessageToTest = {
   initialMessage: ['feat(new-service): Finally solved that problem!'],
   expectedMessage: 'feat(new-service): [JIRA-4321]. Finally solved that problem!',
+  config: {
+    isConventionalCommit: true,
+  },
 };
 
 const firstLineWithCommentMessage: CommitMessageToTest = {
-  initialMessage: ['# This line is comment', 'chore(deps): Finally solved that problem!'],
-  expectedMessage: 'chore(deps): [JIRA-4321]. Finally solved that problem!',
+  initialMessage: ['# This line is comment', 'chore(fix): Finally solved that problem!'],
+  expectedMessage: 'chore(fix): [JIRA-4321]. Finally solved that problem!',
+  config: {
+    isConventionalCommit: true,
+  },
 };
 
 const imitateVerboseCommit: CommitMessageToTest = {
@@ -39,20 +50,69 @@ const imitateVerboseCommit: CommitMessageToTest = {
     '# ------------------------ >8 ------------------------',
   ],
   expectedMessage: '[JIRA-4321].',
+  config: {
+    allowEmptyCommitMessage: true,
+    isConventionalCommit: false,
+  }
 };
 
 const conventionalCommitIncludesTicket = {
   initialMessage: ['feat: [JIRA-4321] Finally solved that problem!'],
   expectedMessage: 'feat: [JIRA-4321] Finally solved that problem!',
+  config: {
+    isConventionalCommit: true,
+  },
 };
 
 const gitRootIsSet = {
-  initialMessage: ['feat: [JIRA-4321] Finally solved that problem!'],
-  expectedMessage: 'feat: [JIRA-4321] Finally solved that problem!',
+  initialMessage: ['feat: Finally solved that problem!'],
+  expectedMessage: 'feat: [JIRA-4321]. Finally solved that problem!',
   config: {
     isConventionalCommit: true,
     messagePattern: '[$J]. $M',
     gitRoot: './.git',
+  },
+};
+
+const allowReplaceAllOccurrencesIsSet = {
+  initialMessage: ['feat(multiple): Finally solved that problem!'],
+  expectedMessage: 'feat(multiple): [JIRA-4321] [Finally solved that problem!]. (JIRA-4321) Finally solved that problem!',
+  config: {
+    allowReplaceAllOccurrences: true,
+    messagePattern: '[$J] [$M]. ($J) $M',
+    isConventionalCommit: true,
+  },
+};
+
+const ignoredBranchesPatternIsSet = {
+  initialMessage: ['feat(ignorebranches): Finally solved that problem!'],
+  expectedMessage: 'feat(ignorebranches): Finally solved that problem!',
+  options: {
+    isBranchIgnored: true,
+  },
+  config: {
+    ignoredBranchesPattern: '^JIRA-4321-test-husky\\d$',
+    isConventionalCommit: true,
+  },
+};
+
+const ignoredBranchesPatternIsSet2 = {
+  initialMessage: ['feat(ignorebranches): Finally solved that problem!'],
+  expectedMessage: 'feat(ignorebranches): [JIRA-4321]. Finally solved that problem!',
+  options: {
+    isBranchIgnored: false,
+  },
+  config: {
+    ignoredBranchesPattern: '^develop$',
+    isConventionalCommit: true,
+  },
+};
+
+const conventionalCommitIncludesMultipleScope = {
+  initialMessage: ['feat(scope1,scope2.scope3, scope4, scope_5): Finally solved that problem!'],
+  expectedMessage: 'feat(scope1,scope2.scope3, scope4, scope_5): [JIRA-4321]. Finally solved that problem!',
+  config: {
+    isConventionalCommit: true,
   },
 };
 
@@ -86,7 +146,13 @@ async function testCommitMessage(
 
   if (commitMessageToTest.config) {
     const pathToConfig = path.join(cwd, '.jirapreparecommitmsgrc');
-    fs.writeFileSync(pathToConfig, JSON.stringify(commitMessageToTest.config));
+    fs.writeFileSync(
+      pathToConfig,
+      JSON.stringify({
+        ...JSON.parse(fs.readFileSync(pathToConfig).toString('utf-8')),
+        ...commitMessageToTest.config,
+      }),
+    );
   }
 
   // Because I can't imitate multiline commit in Windows CLI, I decided to use file
@@ -98,10 +164,16 @@ async function testCommitMessage(
 
   const stdout = await exec('git log', cwd, t);
   const index = stdout.search(/(\[[A-Z]+-\d+])/i);
-  t.is(index > -1, true, `Expected message: ${commitMessageToTest.expectedMessage}`);
-  const index2 = stdout.includes(commitMessageToTest.expectedMessage);
-  t.is(index2, true, `Expected message: ${commitMessageToTest.expectedMessage}`);
+  if (commitMessageToTest.options?.isBranchIgnored) {
+    t.is(index > -1, false, `Log contains JIRA ticket, but shouldn't. Expected message: ${commitMessageToTest.expectedMessage}`);
+  } else {
+    t.is(index > -1, true, `Log doesn't contain JIRA ticket. Expected message: ${commitMessageToTest.expectedMessage}`);
+  }
 
+  const index2 = stdout.includes(commitMessageToTest.expectedMessage);
+  t.is(index2, true, `Log doesn't contain correct message. Expected message: ${commitMessageToTest.expectedMessage}`);
+
+  await exec(`git checkout -- ${path.join(cwd, '.jirapreparecommitmsgrc')}`, __dirname, t);
   await exec(`git update-ref -d HEAD`, cwd, t);
 }
 
@@ -112,6 +184,10 @@ test('husky2 JIRA ticket ID should be in commit message', async (t: ExecutionCon
   await testCommitMessage(imitateVerboseCommit, 'husky2', t);
   await testCommitMessage(conventionalCommitIncludesTicket, 'husky2', t);
   await testCommitMessage(gitRootIsSet, 'husky2', t);
+  await testCommitMessage(allowReplaceAllOccurrencesIsSet, 'husky2', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet, 'husky2', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet2, 'husky2', t);
+  await testCommitMessage(conventionalCommitIncludesMultipleScope, 'husky2', t);
 });
 
 test('husky3 JIRA ticket ID should be in commit message', async (t: ExecutionContext) => {
@@ -121,6 +197,10 @@ test('husky3 JIRA ticket ID should be in commit message', async (t: ExecutionCon
   await testCommitMessage(imitateVerboseCommit, 'husky3', t);
   await testCommitMessage(conventionalCommitIncludesTicket, 'husky3', t);
   await testCommitMessage(gitRootIsSet, 'husky3', t);
+  await testCommitMessage(allowReplaceAllOccurrencesIsSet, 'husky3', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet, 'husky3', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet2, 'husky3', t);
+  await testCommitMessage(conventionalCommitIncludesMultipleScope, 'husky3', t);
 });
 
 test('husky4 JIRA ticket ID should be in commit message', async (t: ExecutionContext) => {
@@ -130,6 +210,10 @@ test('husky4 JIRA ticket ID should be in commit message', async (t: ExecutionCon
   await testCommitMessage(imitateVerboseCommit, 'husky4', t);
   await testCommitMessage(conventionalCommitIncludesTicket, 'husky4', t);
   await testCommitMessage(gitRootIsSet, 'husky4', t);
+  await testCommitMessage(allowReplaceAllOccurrencesIsSet, 'husky4', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet, 'husky4', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet2, 'husky4', t);
+  await testCommitMessage(conventionalCommitIncludesMultipleScope, 'husky4', t);
 });
 
 test('husky5 JIRA ticket ID should be in commit message', async (t: ExecutionContext) => {
@@ -139,4 +223,8 @@ test('husky5 JIRA ticket ID should be in commit message', async (t: ExecutionCon
   await testCommitMessage(imitateVerboseCommit, 'husky5', t);
   await testCommitMessage(conventionalCommitIncludesTicket, 'husky5', t);
   await testCommitMessage(gitRootIsSet, 'husky5', t);
+  await testCommitMessage(allowReplaceAllOccurrencesIsSet, 'husky5', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet, 'husky5', t);
+  await testCommitMessage(ignoredBranchesPatternIsSet2, 'husky5', t);
+  await testCommitMessage(conventionalCommitIncludesMultipleScope, 'husky5', t);
 });